Create the basic structure of the website and add our first recipe
I have created the homepage for our website odin-recipes.
For now it only contains one recipe name and a link to a webpage
describing that recipe.This webpage itself containing the description
and the image of that dish.And also the ingredients and the steps
needed to make the dish.